S05P13C103-61 [BE] :feat 메인페이지에서 라이브리스트 리턴

+++ backend/src/main/java/com/ssafy/api/request/LiveTitlePatchReq.java
@@ -15,7 +15,7 @@ import javax.persistence.Column;
 @ApiModel("LiveUpdate_titlePatchRequest")
 public class LiveTitlePatchReq {
 	@ApiModelProperty(name="방송 PK")
-	int livepk;
+	Long livepk;
 	@ApiModelProperty(name="방송 title")
 	String livetitle;
 }
